requirements.txt
PyYAML
PySimpleGUI
mysql-connector-python